pyproject.toml
[project]
name = "mmse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bleach>=6.3.0",
    "deepagents>=0.4.8",
    "fastapi>=0.135.1",
    "fhirclient>=4.4.0",
    "gradio>=6.9.0",
    "langchain>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-neo4j==0.8.0",
    "langchain-nvidia-ai-endpoints>=1.2.0",
    "langfuse>=4.0.0",
    "langgraph>=1.0.0",
    "neo4j-agent-memory>=0.0.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
